backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai-whisper==20231117
librosa==0.10.2.post1
dtaidistance==2.3.11
ffmpeg-python==0.2.0
python-multipart==0.0.9
pydantic==2.8.2
torch==2.3.0
